On IRIX 6.5, the GNU compiler complained about a `comparison between signed and unsigned' in daemon.c:121. According to the IRIX 6.5 man page (on the SGI website), fileno() should return an `int', which would match the declaration of `fd'. The error message is thus a bit strange.
I have now spent an explicit cast to int to the fileno() call. Maybe
that helps.

+++ pathfinder/compiler/daemon/daemon.c
@@ -118,7 +118,7 @@ daemonize (void)
 in_background:
     /* close all file descs but stderr (log output) */
     for (fd = 0; fd < FOPEN_MAX; fd++)
-        if (fd != fileno (stderr))
+        if (fd != (int) fileno (stderr))
             close (fd);
 
     /* clear a possible EBADF from a close */
